% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cruts.R
\docType{package}
\name{cruts-package}
\alias{cruts}
\alias{cruts-package}
\title{cruts}
\format{\preformatted{ logi NA
}}
\usage{
cruts
}
\description{
Read data from CRU TS3.21: Climatic Research Unit (CRU) Time-Series (TS) Version 3.21 of High Resolution Gridded Data of Month-by-month \url{http://catalogue.ceda.ac.uk/uuid/ac4ecbd554d0dd52a9b575d9666dc42d}
}
\details{
\tabular{ll}{
Package: \tab cruts\cr
Version: \tab 0.1\cr
Date: \tab 2015-07-03\cr
License: \tab GPL-3 \cr
}



section{Dependencies}{
The package \code{cruts} depends upon some other important contributions to CRAN in order to operate; their uses here are indicated:\cr\cr
    sp, raster, stringr, lubridate, ncdf.
}

section{Citation}{
X
}

references{
X
}
}
\author{
Benjamin Taylor, Department of Medicine, Lancaster University
}
\keyword{package}

