% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruts.R
\docType{package}
\name{cruts-package}
\alias{cruts-package}
\alias{cruts}
\title{cruts}
\format{An object of class \code{logical} of length 1.}
\usage{
cruts
}
\description{
Read data from CRU TS3.21: Climatic Research Unit (CRU) Time-Series (TS) Version 3.21 of High Resolution Gridded Data of Month-by-month \url{http://catalogue.ceda.ac.uk/uuid/ac4ecbd554d0dd52a9b575d9666dc42d}
}
\details{
\packageDESCRIPTION{cruts}
\packageIndices{cruts}



section{Dependencies}{
The package \code{cruts} depends upon some other important contributions to CRAN in order to operate; their uses here are indicated:\cr\cr
    sp, raster, stringr, lubridate, ncdf.
}

section{Citation}{
X
}

references{
X
}
}
\author{
Benjamin Taylor, Department of Medicine, Lancaster University
}
\keyword{package}
